Create a generic TransformableOps for adding/removing Sierra records
